+++ CHANGELOG.md
@@ -1,5 +1,65 @@
 # @atomicsmash/launchpad
 
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- ee290b4: Fix 3
+
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- 2e37f1c: Fix 2
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- 319a5e4: Attempt to fix release process
+
+## 1.0.0-beta.1
+
+### Major Changes
+
+- 16c4cf1: Change assets to return array of information instead of src string
+- 16c4cf1: Remove dom ready util, convert to use wp-dom-ready directly
+- 16c4cf1: Set the correct CSS manipulation steps for the theme and plugin.
+- 1c75e65: Made changes to reset styles
+- d92aec6: Update WordPress minimum version
+- defd6eb: Remove styles and fonts from theme.json
+- 12f0623: Added PSR4 autoloading
+
+### Minor Changes
+
+- be1941e: Added launchpad and spacers block pattern categories
+- 37c0ed2: Add typography formatter
+- 12f0623: Added ACF lite constant
+  Added mailtrap on staging/ local
+  Added wp sanitisation allow list
+- de263c0: Added global banner block
+  Added global banner output block
+  Added global banner custom post type
+  Added ACF Pro plugin
+  Added ACF options for global banner
+- 57b4dd2: Added search base template
+- 16c4cf1: Add theme icons to icon block library
+- 721288d: Split editor-styles.scss into two separate scss files, editor-content and editor-controls
+  Adjusted enqueue-assets.php to enqueue these files appropriately
+  Adjusted webpack.mix.cjs to correctly compile these files
+- 0539613: Added spacer pattern
+  Added negative spacer pattern
+- dee57b5: Added 404 base template
+  Added hidden-404 pattern
+- b45dddf: Added archive base template
+- 3cbae73: Added fluid typographies
+  Updated visual comparison screenshots
+- bca069e: Added CTA pattern
+
+### Patch Changes
+
+- 73c8926: Organise functions file into subfiles
+
 ## 1.0.0-beta.0
 
 ### Major Changes